feat!(refactor): remove 'Browser' and 'BrowserProperties', use primary-secondary strategy, fullscreen on monitor.
The window is set to fullscreen on monitor by default in order to make
webview present on target monitor, and not arranged by compositors.

The primary-secondary strategy is used to handle broadcast request. Most
interactive event happens in the primary webview, and secondary webviews
often receive the requests from the primary. Such as changing
background, the primary requests change background, then the primary and
secondaries should all change the background.